requirements.txt
allennlp==0.9.0
beautifulsoup4==4.8.2
django==3.0.5
djangorestframework==3.11.0
django-cors-headers==3.2.1
docopt==0.6.2
langdetect==1.0.7
nltk==3.4.5
overrides==2.8.0
opencv-python==4.2.0.34
Pillow==7.0.0
ptvsd==4.3.2
pudb==2019.2
git+https://github.com/salaniz/pycocoevalcap@a6b79b0fbd0f4f293634831e4e09405de2abc955
pymongo==3.10.1
pyzmq==19.0.0
gunicorn==20.0.4
schema==0.7.1
spacy==2.1.9
stop-words==2018.7.23
tensorboard==2.1.1
textstat==0.6.0
termcolor==1.1.0
tqdm==4.43.0
transformers==2.5.1

torchvision==0.6.1
autopep8==1.4.4
cython==0.29.15
ipykernel==5.1.4
matplotlib==3.1.3
mypy==0.750
numpy==1.17.2
# pillow==7.1.2
pyyaml==5.1.2
pandas